support non-fp32 for colqwen series

+++ src/optimum/rbln/transformers/models/colqwen2/modeling_colqwen2.py
@@ -113,8 +113,7 @@ class RBLNColQwen2ForRetrieval(RBLNModel):
         model.vlm.model.lm_head = model.embedding_proj_layer
         model.vlm.model.config.embedding_dim = model.config.embedding_dim
 
-        # Some of the model weights are different from the model.dtype(vidore/colqwen2-v1.0-hf)
-        return model.to(model.dtype)
+        return model
 
     def forward(
         self,
